Move BlockPayloadHash into separate module
